Rename `internal.rs` to `signer.rs`.
