+++ src/commons/crypto/signing/signers/pkcs11/mod.rs
@@ -27,7 +27,7 @@ pub mod context;
 /// a locally running PyKMIP Python process, and in cases such as the AWS CloudHSM PKCS#11 library is actually
 /// representative of the remote cloud server or server cluster nature of the backend being communicated with by the
 /// PKCS#11 library.
-pub mod internal;
+pub mod signer;
 pub mod session;
 
-pub use internal::Pkcs11Signer;
+pub use signer::Pkcs11Signer;
